Extract all classes from Models.cs
