+++ src/Private/Models/Protocol.cs
@@ -1,0 +1,8 @@
+namespace RepoManager
+{
+    public enum Protocol
+    {
+        HTTPS,
+        SSH
+    }
+}
