+++ src/Private/Models/Provider.cs
@@ -1,0 +1,9 @@
+namespace RepoManager
+{
+    public enum Provider
+    {
+        GitHub,
+        GitLab,
+        BitBucket
+    }
+}
