+++ src/Private/Models/Remote.cs
@@ -1,0 +1,11 @@
+namespace RepoManager
+{
+    public struct Remote
+    {
+        public string Push { get; set; }
+
+        public string Fetch { get; set; }
+
+        public override string ToString() => Fetch.Equals(Push) ? Push : $"{Push} ({Fetch})";
+    }
+}
