+++ src/Private/Models/Scope.cs
@@ -1,0 +1,9 @@
+namespace RepoManager
+{
+    internal enum Scope
+    {
+        Local,
+        Global,
+        System
+    }
+}
